% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_tibble.R
\name{create_plot_tibble}
\alias{create_plot_tibble}
\title{Generate a tibble of data needed for plotting}
\usage{
create_plot_tibble(LR, LR_interval, prior_odds, prior_odds_interval)
}
\arguments{
\item{LR}{Likelihood ratio}

\item{LR_interval}{Uncertainty interval for likelihood ratio}

\item{prior_odds}{Prior odds estimate}

\item{prior_odds_interval}{Uncertainty interval for prior odds estimate}
}
\value{
tibble with data needed for plotting
}
\description{
Generate a tibble of data needed for plotting
}
