% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_breaks.R
\name{get_breaks}
\alias{get_breaks}
\title{Create breaks for likelihood ratio plot}
\usage{
get_breaks(LR, LR_interval, prior_odds, prior_odds_interval)
}
\arguments{
\item{LR}{Likelihood ratio}

\item{LR_interval}{Uncertainty interval for likelihood ratio}

\item{prior_odds}{Prior odds estimate}

\item{prior_odds_interval}{Uncertainty interval for prior odds estimate}
}
\value{
list with breaks for x axis and y axis
}
\description{
Create breaks for likelihood ratio plot
}
