% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labels.R
\name{get_labels}
\alias{get_labels}
\title{Create labels for likelihood ratio plot}
\usage{
get_labels(breaks, digits = 2)
}
\arguments{
\item{breaks}{List generated from get_breaks function}

\item{digits}{Number of significant digits for labels (default, 2)}
}
\value{
list with labels for x axis and y axis
}
\description{
Create labels for likelihood ratio plot
}
