% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_title.R
\name{get_plot_title}
\alias{get_plot_title}
\title{Get title for plot}
\usage{
get_plot_title(LR, LR_interval, prior_odds, prior_odds_interval)
}
\arguments{
\item{LR}{Likelihood ratio}

\item{LR_interval}{Uncertainty interval for likelihood ratio}

\item{prior_odds}{Prior odds estimate}

\item{prior_odds_interval}{Uncertainty interval for prior odds estimate}
}
\value{
title string
}
\description{
Get title for plot
}
