% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_ratio_plot.R
\name{lik_ratio_plot}
\alias{lik_ratio_plot}
\title{Generate likelihood ratio plot}
\usage{
lik_ratio_plot(
  LR,
  LR_interval,
  prior_odds,
  prior_odds_interval,
  plot_rounding_digits = 2
)
}
\arguments{
\item{LR}{Likelihood ratio}

\item{LR_interval}{Uncertainty interval for likelihood ratio}

\item{prior_odds}{Prior odds estimate}

\item{prior_odds_interval}{Uncertainty interval for prior odds estimate}

\item{plot_rounding_digits}{Number of digits after decimal for rounding on plot (default, 2)}
}
\value{
Likelihood ratio plot
}
\description{
Generate likelihood ratio plot
}
\examples{
lik_ratio_plot(10, c(8, 12), .8, c(.6, 1))
}
